Implement UIController for managing currency conversion UI and refactor currency fetching logic
