+++ scripts/CurrencyService.mjs
@@ -30,6 +30,20 @@ class CurrencyService {
     }
 
     /**
+     * Fetches or retrieves currencies from localStorage
+     * @returns {Promise<Array>} - List of currencies
+     */
+    static async getOrFetchCurrencies() {
+        // Check if currencies are already in localStorage
+        let currencies = JSON.parse(localStorage.getItem('currencies'));
+        if (!currencies) {
+            currencies = await this.getCurrencies();
+            localStorage.setItem('currencies', JSON.stringify(currencies));
+        }
+        return currencies;
+    }
+
+    /**
      * Fetches the exchange rate for a given currency and date
      * @param {string} currency - The currency code (e.g., 'USD', 'EUR')
      * @param {string} date - The date in 'YYYY-MM-DD' format
@@ -47,7 +61,8 @@ class CurrencyService {
             const response = await fetch(url);
             if (!response.ok) {
                 const errorData = await response.json();
-                throw new Error(errorData.message);
+                const errorMessage = errorData.message || 'An unknown error occurred while fetching the exchange rate.';
+                throw new Error(errorMessage);
             }
             const data = await response.json();
